<Master Merge>2차 배포

See merge request s08-webmobile2-sub2/S08P12A308!105

+++ Backend/src/main/java/com/ssafy/backend/post/repository/CommentLikeRepository.java
@@ -8,6 +8,6 @@ public interface CommentLikeRepository extends JpaRepository<CommentLike, Long> 
 
     void deleteByCommentIdAndMemberId(long commentId, long memberId);
 
-    Long countCommentLikeByCommentId(long commentId);
+    Integer countCommentLikeByCommentId(long commentId);
 
 }
